requirements.txt
numpy>=1.24.0
scipy>=1.11.0
soundfile>=0.12.1
pyyaml>=6.0.1
mido>=1.3.0
matplotlib>=3.8.0
pyaudio>=0.2.13
tqdm>=4.66.1